pyproject.toml
[tool.poetry]
name = "vms_client"
version = "0.1.3"
description = "VMS client for Exodus Intelligence's API."
authors = ["DevTeam <dev@exodusintel.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "8.0.1"
PyNaCl = "1.4.0"
python-dateutil = "2.8.2"
requests = "2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"
black = "^21.11b1"
isort = "^5.10.1"
flake8-bandit = "^2.1.2"
flake8-bugbear = "^21.11.29"
flake8-builtins = "^1.5.3"
darglint = "^1.8.1"
flake8-docstrings = "^1.6.0"
pre-commit = "^2.17.0"
coverage = "^6.3"
requests-mock = "^1.9.3"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
